Refactor MaskedConvolution2D
Remove mask from persistent parameters as those are subject
to serialization. It's not needed and causes deserialization problems
on CPU (due to numpy `writeable` restrictions). Override `to_gpu`
method to move mask to gpu when needed.

Good option is to add `to_cpu` method as well but I don't need it
at a time.